Remove Which, use package.scala for extension functions
